Updated the Quick Sort with Lomuto partition scheme
I updated the previous implementation of the quick sort algorithm that
uses the Lomuto partition scheme with some clearer function names and
refactored a few things to ensure we are not repeating code as much as
possible. This was done at the time I was writing the solution for the
quick sort algorithm that uses the Hoare partition scheme.